Set push.default to upstream
The default changed from matching to simple in git 2.0, but simple was
only introduced in git 1.7.11, and we have older gits, so I can't just
set it to that. upstream is the closest in functionality, and I think I
don't need the extra security of simple, so I'll just set it
unconditionally to upstream instead of checking the git version.